add get emailCampaignStatistics functionality + tests

+++ src/main/resources/db/migration/V1.01.04__create_campaignStatistics_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS campaign_statistics (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     campaign_id BIGINT NOT NULL REFERENCES email_campaign(id),
-    emailsSent INT,
-    emailsDelivered INT,
-    emailsOpened INT,
-    emailsClicked INT
+    emailsSent BIGINT,
+    emailsDelivered BIGINT,
+    emailsOpened BIGINT,
+    emailsClicked BIGINT
 );
